Add ts_release_date to vintage insert temp table
The query builder for inserting TS will include the release date
in the tuples to be inserted (..., ts_release_date) so the column
must be present in the temporary table. Since release dates are not
needed for vintage series, the date will not be inserted into the final
vintages table though.

See #40