Fix issue causing years with months 2-13
In indexToDate, line 15, floor(zoolike-index) is used to determine
the year of the index. Due to numerical imprecisions, the year 1988 e.g.
may be internally represented as 1987.9999...7726, leading to the wrong
result when flooring. This in turn led to said year having
a 13th month and the next lacking the first.