move release_validity to timeseries_maib
unless the set of keys stored under a given release name
is always exactly the same, storing a single key (or some
totally unrelated ones!) can break previously stored range
info. it is safer to keep these on a per-row level.

Another approach would have been to create a release id and
store both ranges in releases. however then if a subset of keys is
stored these would have to first be removed from the previous